// frontend/authoring/src/scripts/modals/narrative/types.ts
/*
 * LO Platform copyright (C) 2007–2025 LO Ventures LLC.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

import { Dispatch, SetStateAction } from 'react';

import { ProjectResponse } from '../../story/NarrativeMultiverse';
import { NodeName, TypeId } from '../../types/asset';

export type ProjectWithHitCount = ProjectResponse & { hits?: number };

export type StateSetter<A> = Dispatch<SetStateAction<A>>;

export type SearchPathElement = {
  name: string;
  typeId: TypeId;
  title?: string;
  href: string;
};

export type SearchHighlights = Record<string, string[]>;

export type SearchWebHit = {
  project: number;
  projectName: string;
  projectCode?: string | null;
  projectType?: string | null;
  projectStatus?: string | null;
  branch: number;
  branchName: string;
  course: string;
  path: SearchPathElement[];
  highlights: SearchHighlights;
  archived: boolean;
  asset?: any;
};

export type PreviewNode = {
  name: NodeName;
  title: string;
};